Remove tags and remark out of listEntry class
